refactor: modernize mounted checks, fix type casting, and adjust variable modifiers across checkout and profile pages

+++ lib/features/cart/presentation/pages/checkout_payment_page.dart
@@ -466,7 +466,7 @@ class _CheckoutPaymentPageState extends State<CheckoutPaymentPage> {
                   padding: EdgeInsets.symmetric(vertical: 12),
                   child: Divider(),
                 ),
-                _buildPriceRow('cart.total'.tr(), finalTotal, isTotal: true),
+                _buildPriceRow('cart.total'.tr(), finalTotal.toDouble(), isTotal: true),
                 const SizedBox(height: 80),
               ],
             ),
